[FLINK-4505] [cluster mngt] Implement TaskManager component's startup
The TaskManagerRunner now contains the startup logic for the TaskManager's components.
